Resolved conflicts, however -n hint is buggy now (not handling spaces correctly

+++ include/WordData.h
@@ -23,7 +23,7 @@ public:
   unsigned int numAsked;
   unsigned int numCorrect;
   float answeringTime;
-  float getPercentCorrect();
-  float getAverageTime();
+  double getPercentCorrect();
+  double getAverageTime();
 };
 #endif // _WORDDATA_H_
